fix: resolve type comparison error in MONO_HLSB hline optimization
Fixed ViperTypeError: comparison of int and uint at line 1074.

Problem: Comparing uint(bit_pos) with int literal 0
  if bit_pos == 0 and i > 0:  # bit_pos is uint, 0 is int

Solution: Cast bit_pos to int for comparison
  if int(bit_pos) == 0 and i > 0:

Also cast loop variable i to uint when used in arithmetic:
  bit_pos = uint((bit_start + uint(i)) & 7)

This ensures type consistency in @viper code.

+++ fb_opt.py
@@ -1070,14 +1070,14 @@ class FrameBufferHybridOptimized:
 
             if c:
                 for i in range(w):
-                    bit_pos = uint((bit_start + i) & 7)
-                    if bit_pos == 0 and i > 0:
+                    bit_pos = uint((bit_start + uint(i)) & 7)
+                    if int(bit_pos) == 0 and i > 0:
                         offset += 1
                     buf[offset] |= (1 << bit_pos)
             else:
                 for i in range(w):
-                    bit_pos = uint((bit_start + i) & 7)
-                    if bit_pos == 0 and i > 0:
+                    bit_pos = uint((bit_start + uint(i)) & 7)
+                    if int(bit_pos) == 0 and i > 0:
                         offset += 1
                     buf[offset] &= uint(~(1 << bit_pos) & 0xFF)
